Fix: abort when no flags file argument is given Closes #1 Fix program abort when no flags file argument is given. Example: quickc test.c
Output:
terminate called after throwing an instance of 'std::logic_error'
  what():  basic_string::_M_construct null not valid
fish: Job 1, '.quickc test.c' terminated by signal SIGABRT (Abort)

After patch:
quickc test.c

Output:
<RUNS CORRECTLY>

If the user was to run quickc without any flags file argument then
the program would have aborted. This patch fixes that bug by checking
if the arguments given is greater or equal to 3 and if it is then it
will read the third argument as the flags file. If the arguments given
is less than 3 then it will set the flags file to the default flags which
is marked as "." in the code.